expose tables through a standard osquery extension
